+++ CHANGELOG.md
@@ -17,6 +17,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
     `granule_groups.group_id` to use `nextval('granule_group_id_seq')`.
   - Added function `findActiveDuplicateGranules` to `@cumulus/db/src/lib/duplicate-detection` to
     return duplicate granules
+- **CUMULUS-4273**
+  - Added a new query string parameter `includeActiveStatus` to the `GranuleSearch` class for whether to include `group_id` and `state` from the `GranuleGroups` table for granules in api search
 
 ### Notable Changes